+++ controls-core/src/commonMain/kotlin/space/kscience/controls/manager/respondMessage.kt
@@ -74,11 +74,11 @@ public suspend fun DeviceHub.respondHubMessage(request: DeviceMessage): List<Dev
     return try {
         val targetName = request.targetDevice
         if (targetName == null) {
-            buildDeviceTree().mapNotNull {
+            devices.mapNotNull {
                 it.value.respondMessage(it.key, request)
             }
         } else {
-            val device = getOrNull(targetName) ?: error("The device with name $targetName not found in $this")
+            val device = devices[targetName] ?: error("The device with name $targetName not found in $this")
             listOfNotNull(device.respondMessage(targetName, request))
         }
     } catch (ex: Exception) {