+++ controls-server/src/jvmMain/kotlin/space/kscience/controls/server/deviceWebServer.kt
@@ -29,17 +29,16 @@ import kotlinx.serialization.json.put
 import space.kscience.controls.api.DeviceMessage
 import space.kscience.controls.api.PropertyGetMessage
 import space.kscience.controls.api.PropertySetMessage
-import space.kscience.controls.api.getOrNull
 import space.kscience.controls.manager.DeviceManager
 import space.kscience.controls.manager.respondHubMessage
 import space.kscience.dataforge.meta.toMeta
 import space.kscience.dataforge.names.Name
 import space.kscience.dataforge.names.asName
+import space.kscience.dataforge.names.get
 import space.kscience.magix.api.MagixEndpoint
 import space.kscience.magix.api.MagixFlowPlugin
 import space.kscience.magix.api.MagixMessage
 import space.kscience.magix.server.magixModule
-
 
 
 private fun Application.deviceServerModule(manager: DeviceManager) {
@@ -100,10 +99,9 @@ public fun Application.deviceManagerModule(
                         h1 {
                             +"Device server dashboard"
                         }
-                        deviceNames.forEach { deviceName ->
-                            val device =
-                                manager.getOrNull(deviceName)
-                                    ?: error("The device with name $deviceName not found in $manager")
+                        deviceNames.forEach { deviceName: String ->
+                            val device = manager.devices[deviceName]
+                                ?: error("The device with name $deviceName not found in $manager")
                             div {
                                 id = deviceName
                                 h2 { +deviceName }